refactor: resolve circular dependency of injection tokens
