+++ src/lib/src/notifier.module.ts
@@ -7,22 +7,7 @@ import { NotifierConfig, NotifierOptions } from './models/notifier-config.model'
 import { NotifierAnimationService } from './services/notifier-animation.service';
 import { NotifierQueueService } from './services/notifier-queue.service';
 import { NotifierService } from './services/notifier.service';
-
-// tslint:disable variable-name
-
-/**
- * Injection Token for notifier options
- */
-export const NotifierOptionsToken: InjectionToken<NotifierOptions>
-	= new InjectionToken<NotifierOptions>( '[angular-notifier] Notifier Options' );
-
-/**
- * Injection Token for notifier configuration
- */
-export const NotifierConfigToken: InjectionToken<NotifierConfig>
-	= new InjectionToken<NotifierConfig>( '[anuglar-notifier] Notifier Config' );
-
-// tslint:enable variable-name
+import { NotifierConfigToken, NotifierOptionsToken } from './notifier-token';
 
 /**
  * Factory for a notifier configuration with custom options
